snapshot: start breaking things
