+++ options.go
@@ -3,14 +3,14 @@ package export
 import (
 	"context"
 
-	id "github.com/whosonfirst/go-whosonfirst-id"
+	"github.com/whosonfirst/go-whosonfirst-id"
 )
 
 type Options struct {
-	IDProvider id.Provider
+	IdProvider id.Provider
 }
 
-func NewDefaultOptions(ctx context.Context) (*Options, error) {
+func DefaultOptions(ctx context.Context) (*Options, error) {
 
 	provider, err := id.NewProvider(ctx)
 
@@ -18,14 +18,9 @@ func NewDefaultOptions(ctx context.Context) (*Options, error) {
 		return nil, err
 	}
 
-	return NewDefaultOptionsWithProvider(ctx, provider)
-}
-
-func NewDefaultOptionsWithProvider(ctx context.Context, provider id.Provider) (*Options, error) {
-
 	opts := &Options{
-		IDProvider: provider,
+		IdProvider: provider,
 	}
 
-	return opts, nil
+	return opts, err
 }
